+++ includes/object-cache.php
@@ -469,8 +469,16 @@ class WP_Object_Cache
 
                 // Load bundled Predis library
                 if (! class_exists('Predis\Client')) {
-                    $plugin_dir = defined('WP_PLUGIN_DIR') ? WP_PLUGIN_DIR : WP_CONTENT_DIR . '/plugins';
-                    require_once $plugin_dir . '/redis-cache/includes/predis/autoload.php';
+                    $predis = sprintf(
+                        '%s/redis-cache/includes/predis/autoload.php',
+                        defined('WP_PLUGIN_DIR') ? WP_PLUGIN_DIR : WP_CONTENT_DIR . '/plugins'
+                    );
+
+                    if (file_exists($predis)) {
+                        require_once $predis;
+                    } else {
+                        throw new Exception('Predis library not found. Re-install Redis Cache plugin or delete object-cache.php.');
+                    }
                 }
 
                 $options = array();